move sanity fetch to util
